Accredivault-e66d153a697651febde42c5b8161e6744cbd8548/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.29.7
Pillow==10.0.1
python-magic==0.4.27
cryptography==42.0.8
python-dotenv==1.0.1